requirements.txt
redis
pyyaml
flask
requests